requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
sentence-transformers
python-dotenv